commit26
Add CommonJS modules to game, form and reviews.
