+++ src/main.js
@@ -1,5 +1,5 @@
 'use strict';
 
-require('./form');
 require('./game');
-require('./reviews');
+require('./form');
+require('./reviews/reviews');
