+++ src/reviews/filters/filter_types.js
@@ -1,0 +1,12 @@
+'use strict';
+
+/** @enum {string} */
+var Filter = {
+  RECENT: 'reviews-recent',
+  GOOD: 'reviews-good',
+  BAD: 'reviews-bad',
+  POPULAR: 'reviews-popular',
+  ALL: 'reviews-all'
+};
+
+module.exports = Filter;
